Change channel/role position and bitrate type to `u32` (#2149)

+++ src/http/client.rs
@@ -1929,7 +1929,7 @@ impl Http {
         &self,
         guild_id: u64,
         role_id: u64,
-        position: u64,
+        position: u32,
         audit_log_reason: Option<&str>,
     ) -> Result<Vec<Role>> {
         let map = json!([{
